reimplement `ScrollableList`
as two separate traits
- `ListViewContainer` for list-view methods
- another trait for scroller methods
