+++ hackernews_tui/src/view/mod.rs
@@ -1,8 +1,8 @@
 mod async_view;
 mod error_view;
 mod fn_view_wrapper;
-mod list_view;
 mod text_view;
+mod traits;
 
 pub mod article_view;
 pub mod comment_view;
